adding layer and changing the structure
