+++ functions/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "wix-careers-scraper",
+  "private": true,
+  "version": "0.1.0",
+  "description": "AWS Lambda Function that Scrapes Wix.com careers",
+  "main": "handler.js",
+  "type": "module",
+  "devDependencies": {
+    "@sparticuz/chromium": "^119.0.0"
+  },
+  "dependencies": {
+    "puppeteer-core": "^21.5.1"
+  }
+}
