+++ layers/chromium/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "ChromiumLayer",
+  "private": true,
+  "version": "1.0.0",
+  "description": "Chromium layer for AWS Lambda",
+  "dependencies": {
+    "@sparticuz/chromium": "^119.0.0"
+  }
+}
