+++ template.yml
@@ -3,13 +3,32 @@ Transform: AWS::Serverless-2016-10-31
 Description: Node.js Scraper Application
 
 Resources:
+  ChromiumLayer:
+    Type: AWS::Serverless::LayerVersion
+    Properties:
+      Description: Chromium with Node.js integration for AWS Lambda
+      ContentUri: layers/chromium
+      CompatibleRuntimes:
+        - &nodejsRuntime nodejs20.x
+      # Chromium doesn't currently have ARM support; see https://github.com/Sparticuz/chromium#can-i-use-arm-or-graviton-instances
+      CompatibleArchitectures:
+        - &chromiumArch x86_64
+      RetentionPolicy: Delete
+    Metadata:
+      BuildMethod: *nodejsRuntime
+      BuildArchitecture: *chromiumArch
+
   ScraperFunction:
     Type: AWS::Serverless::Function
     Properties:
-      Handler: handler.getWixJobs
-      Runtime: nodejs16.x
-      CodeUri: .
-      MemorySize: 1024
+      Handler: functions/handler.getWixJobs
+      Runtime: *nodejsRuntime
+      Architectures:
+        - *chromiumArch
+      Layers:
+        - !Ref ChromiumLayer
+      # Adjust as necessary
       Timeout: 30
+      MemorySize: 1024
       Policies:
         - AWSLambdaBasicExecutionRole
