+++ test.js
@@ -1,4 +1,4 @@
-import { extractJobs } from './scraper.js';
+import { extractJobs } from './functions/scraper.js';
 
 extractJobs().then(jobs => console.log(jobs));
